requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangChain and LangGraph - Updated compatible versions
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40

# Web research - Updated version
tavily-python>=0.3.0

# Image generation
google-genai>=0.2.0

# Streamlit UI
streamlit>=1.31.0
pandas>=2.1.0

# Additional dependencies
requests>=2.31.0
tiktoken>=0.5.2
